@@ -1,74 +1,25 @@
-import Vue from 'vue';
-import VueRouter, { RouteConfig } from 'vue-router';
+import { createRouter, createWebHistory, RouteRecordRaw } from "vue-router";
+import HomeView from "../views/HomeView.vue";
 
-import BoardList from '../views/board/BoardList.vue';
-import BoardWrite from '../views/board/BoardWrite.vue';
-import BoardDetail from '../views/board/BoardDetail.vue';
-import PageLayout from '../layout/PageLayout.vue';
-import ContentLayout from '../layout/ContentLayout.vue';
-
-Vue.use(VueRouter);
-
-const routes: Array<RouteConfig> = [
+const routes: Array<RouteRecordRaw> = [
   {
-    path: '/',
-    name: 'PageLayout',
-    component: PageLayout,
-    redirect: '/login',
-    children: [
-      {
-        path: 'board',
-        name: 'Board',
-        component: ContentLayout,
-        redirect: '/board/list',
-        children: [
-          {
-            path: 'list',
-            name: 'BoardList',
-            component: BoardList,
-          },
-          {
-            path: 'write',
-            name: 'BoardWrite',
-            component: BoardWrite,
-          },
-          {
-            path: 'update/:idx',
-            name: 'BoardUpdate',
-            component: BoardWrite,
-          },
-          {
-            path: 'detail/:idx',
-            name: 'BoardDetail',
-            component: BoardDetail,
-          },
-        ],
-      },
-    ],
+    path: "/",
+    name: "home",
+    component: HomeView,
   },
-  // {
-  //   path: '/about',
-  //   name: 'About',
-  //   // route level code-splitting
-  //   // this generates a separate chunk (about.[hash].js) for this route
-  //   // which is lazy-loaded when the route is visited.
-  //   component: () =>
-  //     import(/* webpackChunkName: "about" */ '../views/Home.vue'),
-  // },
   {
-    path: '/login',
-    name: 'Login',
+    path: "/about",
+    name: "about",
     // route level code-splitting
     // this generates a separate chunk (about.[hash].js) for this route
     // which is lazy-loaded when the route is visited.
     component: () =>
-      import(/* webpackChunkName: "about" */ '../views/user/Login.vue'),
+      import(/* webpackChunkName: "about" */ "../views/AboutView.vue"),
   },
 ];
 
-const router = new VueRouter({
-  mode: 'history',
-  base: process.env.BASE_URL,
+const router = createRouter({
+  history: createWebHistory(process.env.BASE_URL),
   routes,
 });
 
